refactor(application): delegate main game flow to GameService
